requirements.txt
telethon==1.40.0
python-dotenv==1.1.1
sqlalchemy==2.0.41